fix: 요청서 금액 타입 수정

+++ pickle-common/src/main/java/com/example/pickle_common/consulting/dto/CreateRequestLetterRequest.java
@@ -19,9 +19,9 @@ public class CreateRequestLetterRequest {
     private int answer2;
     private int answer3;
     private int answer4;
-    private int availableInvestAmount;
-    private int desiredInvestAmount;
-    private int monthlyIncome;
+    private long availableInvestAmount;
+    private long desiredInvestAmount;
+    private long monthlyIncome;
 
     private CustomerInfo customerInfo;
     private PbInfo pbInfo;
